requirements.txt
Django==4.0.2
django-cors-headers==3.11.0
django-requestlogs==0.3.0  
djangorestframework==3.13.1
idna==3.3
mysqlclient==2.1.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8